requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cmake==3.26.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.0.7
django-mssql-backend==2.8.1
django-phonenumber-field==7.1.0
django-twilio==0.14.3.2
djangorestframework==3.13.1
dlib @ file:///C:/Users/ASUS/Desktop/face%20Recognisation/Abhinandan_da_machine/Face%20Anirudha/dlib-19.22.99-cp38-cp38-win_amd64.whl#sha256=b02321413b8765159ae552967ab9781a981324b2a389924bdb8c4ccab227f160
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
future==0.18.3
idna==3.3
importlib-resources==5.12.0
iso8601==2.0.0
kiwisolver==1.4.4
Levenshtein==0.21.1
matplotlib==3.7.1
mediapipe==0.10.1
mssql-django==1.1.3
multidict==6.0.4
numpy==1.23.1
openai==0.27.8
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==21.3
pandas==1.4.3
phonenumbers==8.13.20
Pillow==9.2.0
platformdirs==3.0.0
protobuf==3.20.3
pycparser==2.21
PyJWT==2.8.0
pyodbc==4.0.34
pyparsing==3.0.9
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.8.2
python-Levenshtein==0.21.1
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
rapidfuzz==3.2.0
requests==2.28.1
requests-toolbelt==0.9.1
schedule==1.1.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.2
tableau-api-lib==0.1.43
tableauserverclient==0.19.0
tqdm==4.65.0
twilio==7.17.0
typeguard==2.13.3
tzdata==2022.2
urllib3==1.26.10
virtualenv==20.19.0
webencodings==0.5.1
yarl==1.9.2
zipp==3.15.0
